refactor: move allowlist check from generator to mem-pool
