+++ crates/block-producer/src/replay_block.rs
@@ -112,7 +112,6 @@ impl ReplayBlock {
                 &block_info,
                 &raw_tx,
                 L2TX_MAX_CYCLES,
-                None,
             )?;
 
             state.apply_run_result(&run_result)?;
